Refactor container's require of ember-runtime.
