+++ packages/ember-runtime/lib/mixins/container_proxy.js
@@ -3,11 +3,9 @@
 @submodule ember-runtime
 */
 import {
-  deprecate,
   Mixin,
   run
 } from 'ember-metal';
-
 
 /**
   ContainerProxyMixin is used to provide public access to specific
@@ -134,32 +132,3 @@ export default Mixin.create({
     }
   }
 });
-
-export function buildFakeContainerWithDeprecations(container) {
-  let fakeContainer = {};
-  let propertyMappings = {
-    lookup: 'lookup',
-    lookupFactory: '_lookupFactory'
-  };
-
-  for (let containerProperty in propertyMappings) {
-    fakeContainer[containerProperty] = buildFakeContainerFunction(container, containerProperty, propertyMappings[containerProperty]);
-  }
-
-  return fakeContainer;
-}
-
-function buildFakeContainerFunction(container, containerProperty, ownerProperty) {
-  return function() {
-    deprecate(
-      `Using the injected \`container\` is deprecated. Please use the \`getOwner\` helper to access the owner of this object and then call \`${ownerProperty}\` instead.`,
-      false,
-      {
-        id: 'ember-application.injected-container',
-        until: '3.0.0',
-        url: 'http://emberjs.com/deprecations/v2.x#toc_injected-container-access'
-      }
-    );
-    return container[containerProperty](...arguments);
-  };
-}
